Nexus audit iteration 3: fix OOB writes, logic bugs, and overflow safety
- dungeon.c: fix combined-index bounds check in structure1f item material
  binding (bound_regular + bound_special vs max_bindings)
- dungeon.c: fix off-by-one heap overflow in 4bpp VDP1 texel decode for
  odd pixel counts — iterate by texel, not by packed byte
- engine.c: fix party-slot vs champion-index confusion in on_death_update_leader
  (passed slot index instead of champion index at both call sites)
- engine.c: fix champion_status[] indexed by champion index instead of
  party slot for poison/status apply and tick
- combat.c: fix wound penalty using only body table — arm/leg wounds now
  correctly use g_wound_table_limbs (DM.BIN 0x036AF6) instead of
  g_wound_table_body (DM.BIN 0x036AF0)
- dmdf_model.c: add material_id < NEXUS_DMDF_MATERIAL_COUNT bounds check
  in parse_texture_section to prevent stack OOB write
- mns.c: uint64_t byte_count to prevent signed int overflow in texture
  pixel data bounds check
- save_load.c: fseek past unread champion data remainder before reading
  world data to prevent file cursor desync on undersized buffers

+++ src/nexus/nexus_v1_mns.c
@@ -184,7 +184,7 @@ int nexus_v1_mns_decode(const uint8_t *data, int data_size,
                 td->pixel_count = td->width * td->height;
                 if (td->pixel_count > 0 && td->image_offset) {
                     uint32_t abs_off = out->text_offset + td->image_offset;
-                    int byte_count = td->pixel_count * 2;
+                    uint64_t byte_count = (uint64_t)td->pixel_count * 2;
                     if ((uint64_t)td->image_offset + (uint64_t)byte_count <=
                             (uint64_t)text_size &&
                         range_in_file(abs_off, (uint64_t)byte_count,
